вынес расчет юриного расписания в другой файл вынес фильтр сообщений в другой файл
todo: подумать над именами функций
